requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiosmtpd==1.4.6
    # via -r requirements.in
atpublic==5.0
    # via aiosmtpd
attrs==24.2.0
    # via aiosmtpd
blinker==1.9.0
    # via flask
bootstrap-flask==2.4.1
    # via -r requirements.in
click==8.1.7
    # via flask
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via -r requirements.in
faker==33.1.0
    # via -r requirements.in
flask==3.1.0
    # via
    #   -r requirements.in
    #   bootstrap-flask
    #   flask-avatars
    #   flask-dropzone
    #   flask-login
    #   flask-mailman
    #   flask-sqlalchemy
    #   flask-wtf
flask-avatars==0.2.3
    # via -r requirements.in
flask-dropzone==2.0.0
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-mailman==1.1.1
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-whooshee
flask-whooshee==0.9.1
    # via -r requirements.in
flask-wtf==1.2.2
    # via -r requirements.in
idna==3.10
    # via email-validator
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.4
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
pillow==11.0.0
    # via
    #   -r requirements.in
    #   flask-avatars
pyjwt==2.10.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via faker
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.36
    # via flask-sqlalchemy
typing-extensions==4.12.2
    # via
    #   faker
    #   sqlalchemy
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
whoosh==2.7.4
    # via flask-whooshee
wtforms==3.2.1
    # via
    #   bootstrap-flask
    #   flask-wtf
